requirements.txt
streamlit==1.28.0
groq==0.3.0
pandas==2.0.3
plotly==5.17.0
PyPDF2==3.0.1
python-docx==1.1.0
nltk==3.8.1
numpy==1.26.4
